clean up footer player component
