+++ src/components/footer-audio-player/footer-audio-player.scss
